// src/api/services/auth/auth.ts
export interface components {
  schemas: {
    ChangePasswordRequest: {
      oldPassword: string;
      newPassword: string;
    };
    Coin: {
      id?: string;
      symbol?: string;
      name?: string;
      image?: string;
      /** Format: double */
      currentPrice?: number | string;
      /** Format: double */
      priceChangePercentage24h?: number | string;
      /** Format: int64 */
      marketCap?: number | string;
      /** Format: int32 */
      marketCapRank?: number | string;
      description?: null | string;
      /** Format: date-time */
      lastUpdated?: string;
    };
    ForgotPasswordRequest: {
      email: string;
    };
    LoginRequest: {
      email: string;
      password: string;
    };
    MarketOverview: {
      /** Format: int32 */
      id?: number | string;
      /** Format: double */
      totalMarketCap?: number | string;
      /** Format: double */
      marketCapChange?: number | string;
      /** Format: double */
      totalVolume?: number | string;
      /** Format: double */
      volumeChange?: number | string;
      /** Format: double */
      btcDominance?: number | string;
      /** Format: double */
      btcPrice?: number | string;
      /** Format: double */
      ethPrice?: number | string;
      /** Format: date-time */
      timestamp?: null | string;
    };
    MarketStats: {
      coinId?: string;
      /** Format: double */
      currentPrice?: number | string;
      /** Format: int64 */
      marketCap?: number | string;
      /** Format: int32 */
      marketCapRank?: number | string;
      /** Format: double */
      totalVolume?: number | string;
      /** Format: double */
      high24h?: number | string;
      /** Format: double */
      low24h?: number | string;
      /** Format: double */
      circulatingSupply?: number | string;
      /** Format: double */
      totalSupply?: number | string;
      /** Format: double */
      maxSupply?: null | number | string;
      /** Format: double */
      priceChangePercentage24h?: number | string;
    };
    NewsArticle: {
      /** Format: int32 */
      id?: number | string;
      title?: string;
      content?: string;
      summary?: string;
      imageUrl?: string;
      source?: string;
      url?: string;
      isFeatured?: boolean;
      /** Format: date-time */
      publishedAt?: string;
    };
    OhlcvPoint: {
      /** Format: int64 */
      time?: number | string;
      /** Format: double */
      open?: number | string;
      /** Format: double */
      high?: number | string;
      /** Format: double */
      low?: number | string;
      /** Format: double */
      close?: number | string;
      /** Format: double */
      volume?: number | string;
    };
    PriceHistory: {
      /** Format: int32 */
      id?: number | string;
      coinId?: string;
      /** Format: double */
      price?: number | string;
      /** Format: date-time */
      timestamp?: string;
    };
    RefreshTokenRequest: {
      accessToken: string;
      refreshToken: string;
    };
    RegisterRequest: {
      username: string;
      email: string;
      password: string;
    };
    ResetPasswordRequest: {
      email: string;
      token: string;
      newPassword: string;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}